Remove calendar preview
