+++ server/app.ts
@@ -16,8 +16,6 @@ import setUpWebSession from './middleware/setUpWebSession'
 
 import routes from './routes'
 
-import calendarPreview from './routes/calendarPreview' // TODO remove
-
 import type { Services } from './services'
 import populateCurrentBooker from './middleware/populateCurrentBooker'
 
@@ -35,7 +33,6 @@ export default function createApp(services: Services): express.Application {
   app.use(setUpWebRequestParsing())
   app.use(setUpStaticResources())
   nunjucksSetup(app, services.applicationInfo)
-  app.use(calendarPreview(services)) // TODO - remove
   app.use(setupGovukOneLogin())
   app.use(populateCurrentBooker(services.bookerService))
   app.use(setUpCsrf())
