--- server/views/pages/bookingJourney/selectDateAndTime.njk
@@ -1,12 +1,0 @@
-<pre>
-  <h1>Select date and time</h1>
-
-  BOOKER
-  {{ booker | dump(2) }}
-
-  BOOKING JOURNEY
-  {{ bookingJourney | dump(2) }}
-
-  AVAILABLE VISIT SESSIONS
-  {{ visitSessions | dump(2) }}
-</pre>
